Created interface methods for Connect services.
I opted for calling the methods ConnectService because methods generally being verbs in dotnet, I did not want people to think these methods were establishing a connection with the catalog.

This is also deviating from the way that `Service` takes its tag parameters as the tag parameter is being deprecated and will be replaced by filter expressions.